pyproject.toml
[project]
name = "platform-org-backend"
version = "0.3.0"
description = "Platform Org backend (Haier / Rendanheyi)"
requires-python = ">=3.12"
dependencies = [
  "django>=5.0,<6.0",
  "djangorestframework>=3.15,<4.0",
  "django-environ>=0.11,<1.0",
  "psycopg[binary]>=3.2,<4.0",
  "dj-database-url>=2.2,<3.0",
  "gunicorn>=22,<23",
  "uvicorn[standard]>=0.30,<1.0",
  "celery>=5.4,<6.0",
  "redis>=5.0,<6.0",
  "django-cors-headers>=4.4,<5.0",
  "drf-spectacular>=0.27,<1.0",
  "pyjwt[crypto]>=2.9,<3.0",
  "requests>=2.31,<3.0",
  "djangorestframework-simplejwt>=5.3,<6.0",
  "httpx>=0.27,<1.0",
  "whitenoise>=6.0,<7.0",
]
